src/main/java/com/mawulidev/testingframework/CalculatorService.java
package com.mawulidev.testingframework;

import org.springframework.stereotype.Service;

@Service
public class CalculatorService {
    public int add(int a, int b) {
        return a + b;
    }

    public int subtract(int a, int b) {
        return a - b;
    }

    public int multiply(int a, int b) {
        return a * b;
    }

    public int divide(int a, int b) throws ArithmeticException {
        if (b == 0) throw new ArithmeticException("Divide by zero");
        return Math.round(a / b);
    }
}